Toggle layout padding with navbar
